Add is_closure/1,2 and is_export/1,2
Adds new bifs which can be used as guards to check subset types
of funs. Specifically, is_closure/1,2 may be used to check if a fun
is a closure, that is, closes around the module version and/or enviroment
it was created in (i.e., `fun() - ok end`). While, is_export/1,2 is the inverse
and checks that a function is not closed and is simple a function export entry
(i.e. `fun M:F/A`).